New changes to use userid instead of emailid and new truncate command

+++ src/prerna/util/JiraHelper.java
@@ -29,37 +29,101 @@ import prerna.security.HttpHelperUtility;
 
 public class JiraHelper {
 
-	public static NounMetadata listIssue(String projectName, String userName) throws JsonMappingException, JsonProcessingException {
+	public static NounMetadata listIssue(String projectName, String userId) throws JsonMappingException, JsonProcessingException {
+		String msg=null;
 		try {
-			String apiKey = getDBDetails(userName);
-			String auth=userName+":"+apiKey;
+			String apiKey = getDBDetails(userId);
+			String username=getUserName(userId);
+			String auth=username+":"+apiKey;
 			String encodeToString = Base64.getEncoder().encodeToString(auth.getBytes());
 			List<String> jiraIssueDetails = new ArrayList<String>();
 			Map<String, String> map = new HashMap<String, String>();
 			map.put("Authorization", "Basic "+encodeToString);
-			String url = JiraConstant.JiraUrl+"/rest/api/2/search?jql=project=" + projectName;
+			String URL=getURLFromDB(userId);
+			String url = URL+"/rest/api/2/search?jql=project=" + projectName;
 			if (projectName == null || projectName.isEmpty()) {
 				String error = "Project Name for listing issues is missing or null";
 				return new NounMetadata(error, PixelDataType.CUSTOM_DATA_STRUCTURE, PixelOperationType.OPERATION);
 			}
-			String response = HttpHelperUtility.getRequest(url, map, null, null, null);
-			ObjectMapper objectMapper = new ObjectMapper();
-			JsonNode root = objectMapper.readTree(response);
-			JsonNode issues = root.path("issues");
-			for (JsonNode issue : issues) {
-				String name = issue.path("id").asText();
-				jiraIssueDetails.add(name);
-			}
+			List<String> checkUserId = checkUserId(userId);
+				if(!checkUserId.contains(userId)) {
+					msg=userId+" is not present in DB";
+					return new NounMetadata(msg, PixelDataType.CUSTOM_DATA_STRUCTURE,
+							PixelOperationType.OPERATION);
+				}
+					String response = HttpHelperUtility.getRequest(url, map, null, null, null);
+					ObjectMapper objectMapper = new ObjectMapper();
+					JsonNode root = objectMapper.readTree(response);
+					JsonNode issues = root.path("issues");
+					for (JsonNode issue : issues) {
+						String name = issue.path("id").asText();
+						jiraIssueDetails.add(name);
+					}
 			return new NounMetadata(jiraIssueDetails, PixelDataType.CUSTOM_DATA_STRUCTURE,
 					PixelOperationType.OPERATION);
+			
 		} catch (Exception e) {
-			String erroMessage = e.getMessage();
-			return new NounMetadata("Error in listing all issues: " + erroMessage, PixelDataType.CUSTOM_DATA_STRUCTURE,
+			msg = e.getMessage();
+			return new NounMetadata("Error in listing all issues: " + msg, PixelDataType.CUSTOM_DATA_STRUCTURE,
 					PixelOperationType.OPERATION);
 		}
 	}
 
-	private static String getDBDetails(String userName) throws Exception, SQLException {
+	private static String getURLFromDB(String userId) {
+		String error=null;
+		try {
+			String URL = null;
+			String tableName = null;
+			IDatabaseEngine database = Utility.getDatabase("c44b138d-aa8e-42cc-a925-6c2ac855df64");
+			List<String> pixelConcepts = database.getPixelConcepts();
+			for (String element : pixelConcepts) {
+				tableName = element;
+			}
+			String getURLQuery = "select URL from " + tableName + " WHERE JIRAPROFILE_UNIQUE_ROW_ID='" + userId + "'";
+			HashMap<String, String> hashmap = (HashMap<String, String>) database.execQuery(getURLQuery);
+			Object string = hashmap.get("RESULTSET_OBJECT");
+			if (string instanceof ResultSet) {
+				ResultSet rs = (ResultSet) string;
+				while (rs.next()) {
+					URL = rs.getString("URL");
+				}
+			}
+			return URL;
+		} catch (Exception e) {
+			error=e.getMessage();
+		}
+		return error;
+	}
+
+	private static String getUserName(String userId) {
+		String error=null;
+		try {
+			String userName = null;
+			String tableName = null;
+			IDatabaseEngine database = Utility.getDatabase("c44b138d-aa8e-42cc-a925-6c2ac855df64");
+			List<String> pixelConcepts = database.getPixelConcepts();
+			for (String element : pixelConcepts) {
+				tableName = element;
+			}
+			String selectQuery = "SELECT UserId FROM " + tableName + " WHERE JIRAPROFILE_UNIQUE_ROW_ID='" + userId
+					+ "'";
+			HashMap<String, String> hashmap = (HashMap<String, String>) database.execQuery(selectQuery);
+			Object string = hashmap.get("RESULTSET_OBJECT");
+			if (string instanceof ResultSet) {
+				ResultSet rs = (ResultSet) string;
+				while (rs.next()) {
+					userName = rs.getString("UserId");
+				}
+			}
+			return userName;
+		} catch (Exception e) {
+			error=e.getMessage();
+			
+		}
+		return error;
+	}
+
+	private static String getDBDetails(String userId) throws Exception, SQLException {
 		String tableName=null;
 		String apiKey=null;
 		IDatabaseEngine database = Utility.getDatabase("c44b138d-aa8e-42cc-a925-6c2ac855df64");
@@ -67,7 +131,7 @@ public class JiraHelper {
 		for (String element : pixelConcepts) {
 			tableName = element;
 		}
-		String insertQuery = "SELECT Apikey FROM " + tableName + " WHERE UserId='" + userName + "'"+" ORDER BY DATE_CREATED DESC LIMIT 1";
+		String insertQuery = "SELECT Apikey FROM " + tableName + " WHERE JIRAPROFILE_UNIQUE_ROW_ID='" + userId + "'";
 		HashMap<String, String> hashmap = (HashMap<String, String>) database.execQuery(insertQuery);
 		Object string = hashmap.get("RESULTSET_OBJECT");
 		if(string instanceof ResultSet) {
@@ -79,10 +143,13 @@ public class JiraHelper {
 		return apiKey;
 	}
 
-	public static NounMetadata createIssue(String summary, String description, String istype, String projectName, String userName) {
+	public static NounMetadata createIssue(String summary, String description, String istype, String projectName, String userId) {
+		String msg=null;
 		try {
-			String apiKey = getDBDetails(userName);
-			String auth=userName+":"+apiKey;
+			Map<String, Object> responseMap =null;
+			String apiKey = getDBDetails(userId);
+			String username=getUserName(userId);
+			String auth=username+":"+apiKey;
 			String encodeToString = Base64.getEncoder().encodeToString(auth.getBytes());
 			Map<String, String> map = new HashMap<String, String>();
 			map.put("Authorization", "Basic "+encodeToString);
@@ -92,10 +159,11 @@ public class JiraHelper {
 				return new NounMetadata(error, PixelDataType.CUSTOM_DATA_STRUCTURE, PixelOperationType.OPERATION);
 			}
 			String nearestNeigborResponse = null;
-			String url = JiraConstant.JiraUrl+"/rest/api/2/issue";
+			String URL=getURLFromDB(userId);
+			String url = URL+"/rest/api/2/issue";
 			Project project = new Project();
 			project.setKey(projectName);
-			IssueType issuetype = new IssueType();
+			IssueType issuetype = new IssueType(); 
 			issuetype.setName(istype);
 			Fields fields = new Fields();
 			fields.setProject(project);
@@ -103,24 +171,35 @@ public class JiraHelper {
 			fields.setDescription(description);
 			fields.setIssuetype(issuetype);
 			JiraRequestBodyModel jiraRequestBodyModel = new JiraRequestBodyModel();
-			jiraRequestBodyModel.setFields(fields);
+			jiraRequestBodyModel.setFields(fields); 
 			Gson gson = new GsonBuilder().setPrettyPrinting().create();
 			String body = gson.toJson(jiraRequestBodyModel);
-			nearestNeigborResponse = HttpHelperUtility.postRequestStringBody(url, map, body,
-					ContentType.APPLICATION_JSON, null, null, null);
-			Map<String, Object> responseMap = gson.fromJson(nearestNeigborResponse, Map.class);
-			return new NounMetadata(responseMap, PixelDataType.CUSTOM_DATA_STRUCTURE, PixelOperationType.OPERATION);
+			List<String> checkUserId = checkUserId(userId);
+				if(!checkUserId.contains(userId)) {
+					msg="User id "+userId+ " is not present in DB";
+					return new NounMetadata(msg, PixelDataType.CUSTOM_DATA_STRUCTURE, PixelOperationType.OPERATION);
+				}
+					nearestNeigborResponse = HttpHelperUtility.postRequestStringBody(url, map, body,
+							ContentType.APPLICATION_JSON, null, null, null);
+					responseMap = gson.fromJson(nearestNeigborResponse, Map.class);
+					Object jiraId = responseMap.get("id");
+					Object link = responseMap.get("self");
+				
+			
+			return new NounMetadata("Jira id: "+jiraId+","+" Jira link: "+link, PixelDataType.CUSTOM_DATA_STRUCTURE, PixelOperationType.OPERATION);
 		} catch (Exception e) {
-			String errorMessage = e.getMessage();
-			return new NounMetadata("Error in creating new issue: " + errorMessage, PixelDataType.CUSTOM_DATA_STRUCTURE,
+			msg= e.getMessage();
+			return new NounMetadata("Error in creating new issue: " + msg, PixelDataType.CUSTOM_DATA_STRUCTURE,
 					PixelOperationType.OPERATION);
 		}
 	}
 
-	public static NounMetadata deleteIssue(String jiraId, String userName) {
+	public static NounMetadata deleteIssue(String jiraId, String userId) {
+		String msg=null;
 		try {
-			String apiKey = getDBDetails(userName);
-			String auth=userName+":"+apiKey;
+			String apiKey = getDBDetails(userId);
+			String username=getUserName(userId);
+			String auth=username+":"+apiKey;
 			String encodeToString = Base64.getEncoder().encodeToString(auth.getBytes());
 			Map<String, String> map = new HashMap<String, String>();
 			map.put("Authorization", "Basic "+encodeToString);
@@ -128,16 +207,88 @@ public class JiraHelper {
 				String error = "Jira id can not be empty or null";
 				return new NounMetadata(error, PixelDataType.CUSTOM_DATA_STRUCTURE, PixelOperationType.OPERATION);
 			}
-			String url = JiraConstant.JiraUrl+"/rest/api/2/issue/" + jiraId;
-			HttpHelperUtility.deleteRequestStringBody(url, map, null, null, null);
-			return new NounMetadata(jiraId + " succesfully deleted", PixelDataType.CUSTOM_DATA_STRUCTURE,
+			String URL=getURLFromDB(userId);
+			String url = URL+"/rest/api/2/issue/" + jiraId;
+			List<String> checkUserId = checkUserId(userId);
+				if(!checkUserId.contains(userId)) {
+					msg="User id "+userId+ " is not present in DB";
+					return new NounMetadata(msg, PixelDataType.CUSTOM_DATA_STRUCTURE,
+							PixelOperationType.OPERATION);
+				}
+					HttpHelperUtility.deleteRequestStringBody(url, map, null, null, null);
+			return new NounMetadata("Jira id "+jiraId + " succesfully deleted", PixelDataType.CUSTOM_DATA_STRUCTURE,
 					PixelOperationType.OPERATION);
 		} catch (Exception e) {
-			String errorMessage = e.getMessage();
-			return new NounMetadata("Error in deleting issue: " + errorMessage, PixelDataType.CUSTOM_DATA_STRUCTURE,
+			msg = e.getMessage();
+			return new NounMetadata("Error in deleting issue: " + msg, PixelDataType.CUSTOM_DATA_STRUCTURE,
 					PixelOperationType.OPERATION);
 		}
 
 	}
+
+	public static NounMetadata truncateData(String userId) {
+		Boolean truncateFlag=false;
+		String msg=null;
+		String error=null;
+		try {
+			String tableName = null;
+			long Uid;
+			IDatabaseEngine database = Utility.getDatabase("c44b138d-aa8e-42cc-a925-6c2ac855df64");
+			List<String> pixelConcepts = database.getPixelConcepts();
+			for (String element : pixelConcepts) {
+				tableName = element;
+			}
+			List<String> checkUserId = checkUserId(userId);
+				if(!checkUserId.contains(userId)) {
+					msg="User id "+userId+ " is not present in DB";
+					return new NounMetadata(msg, PixelDataType.CUSTOM_DATA_STRUCTURE,
+							PixelOperationType.OPERATION);
+				}
+					String truncateQuery = "Delete from JIRAPROFILE";
+					database.removeData(truncateQuery);
+					truncateFlag=true;
+					Uid=Long.valueOf(userId);
+					if(truncateFlag==true) {
+						msg="Table truncated succesfully by user with user id: "+Uid;
+						
+					}
+			return new NounMetadata(msg, PixelDataType.CUSTOM_DATA_STRUCTURE,
+					PixelOperationType.OPERATION);
+			
+		} catch (Exception e) {
+			error=e.getMessage();
+		}
+		return new NounMetadata("Data not truncated with error message: "+error, PixelDataType.CUSTOM_DATA_STRUCTURE,
+				PixelOperationType.OPERATION);
+	}
+
+	private static List<String> checkUserId(String userId) {
+		List<String> userIds = new ArrayList<String>();
+		try {
+			String tableName = null;
+			String userID;
+			IDatabaseEngine database = Utility.getDatabase("c44b138d-aa8e-42cc-a925-6c2ac855df64");
+			List<String> pixelConcepts = database.getPixelConcepts();
+			for (String element : pixelConcepts) {
+				tableName = element;
+			}
+			String query = " SELECT JIRAPROFILE_UNIQUE_ROW_ID from " + tableName;
+			HashMap<String, String> hashmap = (HashMap<String, String>) database.execQuery(query); 
+			Object string = hashmap.get("RESULTSET_OBJECT");
+			if (string instanceof ResultSet) {
+				ResultSet rs = (ResultSet) string;
+				while (rs.next()) {
+					userID = rs.getString("JIRAPROFILE_UNIQUE_ROW_ID");
+					userIds.add(userID);
+				}
+			}
+			return userIds;
+		} catch (Exception e) {
+			e.printStackTrace();
+			return userIds;
+		}
+	}
+
+	
 
 }
